Fixed local fixture test to test movements
